Optimisations (#59)
* small fixes for status name

* ignore and ignored has been fixed added link to online documentation in help
updated logo
fixed some minor errors

* a lot small fixes

* updated date variables to work even when outside of voice channel

* fixed problem with high database request when nothing is happening

* updated logo

* updated focus functionality to work 100% and added infinite time
focus channels are now voice channels and can be updated with set
update focus channel in correct place in code

* updated focud to work properly
removed spotify
improved code quality and clarity

* added two logo spinners

* covid and crypto message updated and rich embedded portal logo image transitioned to gif spinner

* spinner gif update

* updated level ranks and whoami to have logo spinner

* update message reply, removed unused arguments
small tweaks

* update portal logo

+++ src/index.ts
@@ -144,13 +144,21 @@ client.on('ready', () =>
 );
 
 // This event triggers when a member joins or leaves a voice channel
-client.on('voiceStateUpdate', (oldState: VoiceState, newState: VoiceState) =>
+client.on('voiceStateUpdate', (oldState: VoiceState, newState: VoiceState) => {
+	const new_channel = newState.channel; // join channel
+	const old_channel = oldState.channel; // left channel
+
+	if ((old_channel && new_channel) &&
+		(new_channel.id === old_channel.id)) {
+		return;
+	}
+
 	event_loader('voiceStateUpdate', {
 		'client': client,
 		'oldState': oldState,
 		'newState': newState
-	})
-);
+	});
+});
 
 // runs on every single message received, from any channel or DM
 client.on('message', async (message: Message) => {
@@ -196,22 +204,19 @@ client.on('message', async (message: Message) => {
 				}
 
 				if (!message.guild) {
-					message_reply(false, message.channel, message, message.author,
-						'could not fetch guild of message', client);
+					message_reply(false, message, message.author, 'could not fetch guild of message');
 					return false;
 				}
 
 				if (!message.member) {
-					message_reply(false, message.channel, message, message.author,
-						'could not fetch member of message', client);
+					message_reply(false, message, message.author, 'could not fetch member of message');
 					return false;
 				}
 
 				fetch_guild_authenticate(message.guild.id, message.member.id)
 					.then(authenticate => {
 						if (!authenticate) {
-							message_reply(false, message.channel, message, message.author,
-								'server is not in database, please contact portal support', client);
+							message_reply(false, message, message.author, 'server is not in database, please contact portal support');
 							return false;
 						}
 
@@ -222,13 +227,16 @@ client.on('message', async (message: Message) => {
 						// Ignore any message that does not start with prefix
 						if (message.content.indexOf(authenticate.prefix) !== 0) {
 							if (message.content === 'prefix') {
-								message_reply(true, message.channel, message, message.author,
-									`portal's prefix is \`${authenticate.prefix}\``, client);
+								message_reply(true, message, message.author,
+									`portal's prefix is \`${authenticate.prefix}\``, false);
+								if (message.deletable) {
+									message.delete();
+								}
 							}
 							return false;
 						}
 
-						// Separate function name, and arguments of function
+						// Separate command name and arguments
 						const args = message.content.slice(authenticate.prefix.length).trim().split(/ +/g);
 
 						const cmd_only = args.shift();
@@ -256,8 +264,8 @@ client.on('message', async (message: Message) => {
 						}
 
 						if (!command_options) {
-							message_reply(false, message.channel, message, message.author,
-								'could not get command option', client);
+							message_reply(false, message, message.author,
+								'could not get command option');
 							return false;
 						}
 
@@ -269,43 +277,42 @@ client.on('message', async (message: Message) => {
 
 						if (command_options.auth && message.member) {
 							if (!is_authorised(authenticate.member_list, authenticate.auth_role, message.member)) {
-								message_reply(false, message.channel, message, message.author,
-									'you are not authorised to use this command', client);
+								message_reply(false, message, message.author, 'you are not authorised to use this command');
 								return false;
 							}
 						}
 
 						if (!message.guild) {
-							message_reply(false, message.channel, message, message.author,
-								'could not fetch guild from message', client);
+							message_reply(false, message, message.author,
+								'could not fetch guild from message');
 							return false;
 						}
 
 						fetch_guild(message.guild.id)
 							.then(guild_object => {
 								if (!guild_object) {
-									message_reply(false, message.channel, message, message.author,
-										'server is not in database, please contact portal support', client);
+									message_reply(false, message, message.author,
+										'server is not in database, please contact portal support');
 									return false;
 								}
 
 								if (!command_options) {
-									message_reply(false, message.channel, message, message.author,
-										'could not get command option', client);
+									message_reply(false, message, message.author,
+										'could not get command option');
 									return false;
 								}
 
 								command_loader(message, cmd, args, type, command_options, path_to_command, guild_object);
 							})
-							.catch(error => {
-								message_reply(false, message.channel, message, message.author,
-									'could not get guild from message', client);
+							.catch(e => {
+								message_reply(false, message, message.author,
+									`could not get guild from message (${e})`);
 								return false;
 							});
 					})
-					.catch(error => {
-						message_reply(false, message.channel, message, message.author,
-							'could not get authentication data from message', client);
+					.catch(e => {
+						message_reply(false, message, message.author,
+							`could not get authentication data from message (${e})`);
 						return false;
 					});
 			}
@@ -320,6 +327,10 @@ function command_loader(
 	message: Message, cmd: string, args: string[], type: string, command_options: CommandOptions,
 	path_to_command: string, guild_object: GuildPrtl
 ): boolean {
+	if (config.debug === true) {
+		console.log(`├─ cmd (${cmd})`);
+	}
+
 	if (type === 'none' && command_options.time === 0) {
 		require(`./commands/${path_to_command}/${cmd}.js`)(message, args, guild_object, client)
 			.then((response: ReturnPormise) => {
@@ -327,8 +338,8 @@ function command_loader(
 					if ((command_options && response.value && response.value !== '') &&
 						(command_options.reply || response.result === false)) {
 						message_reply(
-							response.result, message.channel, message, message.author, response.value,
-							client, command_options ? command_options.auto_delete : true
+							response.result, message, message.author, response.value,
+							command_options ? command_options.auto_delete : true
 						);
 					}
 				}
@@ -336,17 +347,18 @@ function command_loader(
 		return true;
 	}
 
-	const active = active_cooldowns[type === 'guild' ? 'guild' : 'member'].find(active_current => {
-		if (active_current.command === cmd) {
-			if (type === 'member' && active_current.member === message.author.id) {
-				return true;
+	const active = active_cooldowns[type === 'guild' ? 'guild' : 'member']
+		.find(active_current => {
+			if (active_current.command === cmd) {
+				if (type === 'member' && active_current.member === message.author.id) {
+					return true;
+				}
+				if (type === 'guild') {
+					return true;
+				}
 			}
-			if (type === 'guild') {
-				return true;
-			}
-		}
-		return false;
-	});
+			return false;
+		});
 
 	if (active) {
 		const time = time_elapsed(active.timestamp, command_options.time);
@@ -354,11 +366,10 @@ function command_loader(
 			? '.*'
 			: `, as it was used again in* **${message.guild?.name}**`;
 
-		message_reply(false, message.channel, message, message.author,
+		message_reply(false, message, message.author,
 			`you need to wait **${pad(time.remaining_min)}:` +
 			`${pad(time.remaining_sec)}/${pad(time.timeout_min)}:` +
-			`${pad(time.timeout_sec)}** *to use* **${cmd}** *again${type_for_msg}`,
-			client);
+			`${pad(time.timeout_sec)}** *to use* **${cmd}** *again${type_for_msg}`);
 
 		return false;
 	}
@@ -382,16 +393,17 @@ function command_loader(
 			}
 
 			if (command_options && response.value && response.value !== '' && (command_options.reply || response.result === false))
-				message_reply(
-					response.result, message.channel, message, message.author, response.value,
-					client, command_options.auto_delete
-				);
+				message_reply(response.result, message, message.author, response.value, command_options.auto_delete);
 		});
 
 	return false;
 }
 
 function event_loader(event: string, args: any): void {
+	if (event === 'presenceUpdate') {
+		return;
+	}
+
 	if (config.debug === true) {
 		console.log(`├─ event (${event})`);
 	}
@@ -455,12 +467,12 @@ async function portal_channel_handler(
 							.then(r => {
 								const reply_message = r ? 'successfully removed from ignored channels'
 									: 'failed to remove from ignored channels'
-								message_reply(true, message.channel, message, message.author,
-									reply_message, client);
+								message_reply(true, message, message.author,
+									reply_message);
 							})
 							.catch(e => {
-								message_reply(false, message.channel, message, message.author,
-									'failed to remove from ignored channels', client);
+								message_reply(false, message, message.author,
+									'failed to remove from ignored channels');
 							});
 					}
 
@@ -470,22 +482,25 @@ async function portal_channel_handler(
 					if (message.content === './url') {
 						remove_url(guild_object.id, message.channel.id)
 							.then(r => {
-								message_reply(true, message.channel, message, message.author,
-									r ? 'successfully removed url channel' : 'failed to remove url channel', client);
+								message_reply(true, message, message.author,
+									r ? 'successfully removed url channel' : 'failed to remove url channel');
 							})
 							.catch(e => {
-								message_reply(false, message.channel, message, message.author,
-									'failed to remove url channel', client);
+								message_reply(false, message, message.author,
+									'failed to remove url channel');
 							});
 					}
 					else if (is_url(message.content)) {
 						client_talk(client, guild_object, 'url');
 					}
 					else {
-						client_talk(client, guild_object, 'read_only');
-						message_reply(false, message.channel, message, message.author,
-							'url-only channel', client);
-						message.delete();
+						// client_talk(client, guild_object, 'read_only');
+						message.author
+							.send(`${message.channel} is a url-only channel`)
+							.catch(console.error);
+						if (message.deletable) {
+							message.delete();
+						}
 					}
 
 					return resolve(true);
@@ -501,28 +516,20 @@ async function portal_channel_handler(
 						const music_data = new MusicData('null', 'null', []);
 						set_music_data(guild_object.id, music_data)
 							.then(r => {
-								message_reply(true, message.channel, message, message.author,
+								message_reply(true, message, message.author,
 									r
 										? 'successfully removed music channel'
-										: 'failed to remove music channel'
-									, client);
+										: 'failed to remove music channel');
 							})
 							.catch(e => {
-								message_reply(false, message.channel, message, message.author,
-									'failed to remove music channel', client);
+								message_reply(false, message, message.author,
+									'failed to remove music channel');
 							});
 					} else {
 						const voice_connection = client.voice?.connections.find(c =>
 							c.channel.guild.id === message.guild?.id);
 
-						if (!message.guild) {
-							if (message.deletable) {
-								message.delete();
-							}
-							return resolve(false);
-						}
-
-						if (!message.member) {
+						if (!message.guild || !message.member) {
 							if (message.deletable) {
 								message.delete();
 							}
@@ -532,8 +539,8 @@ async function portal_channel_handler(
 						start(voice_connection, client, message.member.user, message.guild, guild_object, message.content)
 							.then(joined => {
 								// message_reply(
-								// 	joined.result, message.channel, message,
-								// 	message.author, joined.value, guild_object, client, true
+								// 	joined.result, message,
+								// 	message.author, joined.value, guild_object, true
 								// );
 
 								if (message.deletable) {
@@ -542,8 +549,8 @@ async function portal_channel_handler(
 							})
 							.catch(error => {
 								// message_reply(
-								// 	false, message.channel, message,
-								// 	message.author, error, guild_object, client, true
+								// 	false, message,
+								// 	message.author, error, guild_object, true
 								// );
 
 								if (message.deletable) {
@@ -566,14 +573,13 @@ function ranking_system(message: Message): void {
 		.then(guild_object => {
 			if (guild_object) {
 				const level = add_points_message(message, guild_object);
-				if (level)
-					message_reply(true, message.channel, message, message.author,
-						`you reached level ${level}!`, client);
+				if (level) {
+					message_reply(true, message, message.author, `you reached level ${level}!`);
+				}
 			}
 		})
 		.catch(e => {
-			message_reply(true, message.channel, message, message.author,
-				'an error occured while accesing data', client);
+			message_reply(true, message, message.author, 'an error occured while accesing data');
 		});
 }
